refactor endpoints
