+++ src/reduct/api/common.h
@@ -3,7 +3,6 @@
 #ifndef REDUCT_STORAGE_HANDLERS_COMMON_H
 #define REDUCT_STORAGE_HANDLERS_COMMON_H
 
-#include <google/protobuf/util/json_util.h>
 #include <uWebSockets/App.h>
 
 #include "reduct/async/loop.h"
@@ -12,19 +11,6 @@
 #include "reduct/core/logger.h"
 
 namespace reduct::api {
-
-template <class T>
-std::string PrintToJson(T &&msg) {
-  using google::protobuf::util::JsonPrintOptions;
-  using google::protobuf::util::MessageToJsonString;
-
-  std::string data;
-  JsonPrintOptions options;
-  options.preserve_proto_field_names = true;
-  options.always_print_primitive_fields = true;
-  MessageToJsonString(msg, &data, options);
-  return data;
-}
 
 template <bool SSL>
 struct AsyncHttpReceiver {
